requirements.txt
llama-index-core>=0.10.0
llama-index-llms-openrouter>=0.1.0